Moved implementation code out side of gen folder
