+++ lib/infrastructure/cbj_comp/cbj_comp_repository.dart
@@ -7,10 +7,10 @@ import 'package:cybear_jinni/domain/create_home/i_create_home_repository.dart';
 import 'package:cybear_jinni/domain/manage_network/i_manage_network_repository.dart';
 import 'package:cybear_jinni/domain/manage_network/manage_network_entity.dart';
 import 'package:cybear_jinni/domain/user/user_entity.dart';
-import 'package:cybear_jinni/infrastructure/core/gen/cbj_app_server/cbj_app_server_d.dart';
+import 'package:cybear_jinni/infrastructure/cbj_app_server/cbj_app_server_d.dart';
 import 'package:cybear_jinni/infrastructure/core/gen/cbj_hub_server/protoc_as_dart/cbj_hub_server.pbgrpc.dart';
 import 'package:cybear_jinni/infrastructure/core/gen/security_bear/client/protoc_as_dart/security_bear_connections.pbgrpc.dart';
-import 'package:cybear_jinni/infrastructure/core/gen/security_bear/client/security_bear_server_client.dart';
+import 'package:cybear_jinni/infrastructure/security_bear_client/security_bear_server_client.dart';
 import 'package:cybear_jinni/injection.dart';
 import 'package:dartz/dartz.dart';
 import 'package:injectable/injectable.dart';
