+++ lib/infrastructure/hub_client/hub_connection_repository.dart
@@ -6,9 +6,9 @@ import 'package:cybear_jinni/domain/hub/hub_failures.dart';
 import 'package:cybear_jinni/domain/hub/hub_value_objects.dart';
 import 'package:cybear_jinni/domain/hub/i_hub_connection_repository.dart';
 import 'package:cybear_jinni/domain/local_db/i_local_db_repository.dart';
-import 'package:cybear_jinni/infrastructure/core/gen/cbj_hub_server/hub_client.dart';
 import 'package:cybear_jinni/infrastructure/core/gen/cbj_hub_server/protoc_as_dart/cbj_hub_server.pb.dart';
-import 'package:cybear_jinni/infrastructure/hub_connection/hub_dtos.dart';
+import 'package:cybear_jinni/infrastructure/hub_client/hub_client.dart';
+import 'package:cybear_jinni/infrastructure/hub_client/hub_dtos.dart';
 import 'package:cybear_jinni/injection.dart';
 import 'package:cybear_jinni/utils.dart';
 import 'package:dartz/dartz.dart';
@@ -24,7 +24,7 @@ import 'package:ping_discover_network_forked/ping_discover_network_forked.dart';
 @LazySingleton(as: IHubConnectionRepository)
 class HubConnectionRepository extends IHubConnectionRepository {
   HubConnectionRepository() {
-    if (currentEnv == Env.prod) {
+    if (currentEnv == Env.dev) {
       hubPort = 60055;
     } else {
       hubPort = 50055;
