+++ lib/infrastructure/hub_client/hub_dtos.dart
