+++ lib/infrastructure/security_bear_client/security_bear_server_client.dart
@@ -2,7 +2,6 @@ import 'dart:async';
 
 import 'package:cybear_jinni/domain/create_home/i_create_home_repository.dart';
 import 'package:cybear_jinni/domain/manage_network/manage_network_entity.dart';
-// import 'package:cybear_jinni/infrastructure/core/constant_credentials.dart';
 import 'package:cybear_jinni/infrastructure/core/gen/security_bear/client/protoc_as_dart/security_bear_connections.pbgrpc.dart';
 import 'package:cybear_jinni/injection.dart';
 import 'package:grpc/grpc.dart';
@@ -46,10 +45,12 @@ class SecurityBearServerClient {
   }
 
   static Future<SBCommendStatus?> setFirebaseAccountInformation(
-      String deviceIp, ManageNetworkEntity secondWifiEntity) async {
+    String deviceIp,
+    ManageNetworkEntity secondWifiEntity,
+  ) async {
     final ClientChannel channel = createSmartServerClient(deviceIp);
     final SecurityBearClient stub = SecurityBearClient(channel);
-    SBCommendStatus responseSBCommendStatus = SBCommendStatus();
+    final SBCommendStatus responseSBCommendStatus = SBCommendStatus();
     //
     // final SBFirebaseAccountInformation sbFirebaseAccountInformation =
     //     SBFirebaseAccountInformation()
@@ -61,15 +62,18 @@ class SecurityBearServerClient {
     try {
       final ManageNetworkEntity manageFirstNetworkEntity =
           (await getIt<ICreateHomeRepository>().getFirstWifi()).getOrElse(
-              () => throw "Can't get current home first WiFi from Firestore");
+        () => throw "Can't get current home first WiFi from Firestore",
+      );
 
       final SecurityBearSetup securityBearSetup = SecurityBearSetup()
         ..wiFiFirstPriority = WiFiInformation(
-            wiFiName: manageFirstNetworkEntity.name!.getOrCrash(),
-            wiFiPassword: manageFirstNetworkEntity.pass!.getOrCrash())
+          wiFiName: manageFirstNetworkEntity.name!.getOrCrash(),
+          wiFiPassword: manageFirstNetworkEntity.pass!.getOrCrash(),
+        )
         ..wiFiSecondPriority = WiFiInformation(
-            wiFiName: secondWifiEntity.name!.getOrCrash(),
-            wiFiPassword: secondWifiEntity.pass!.getOrCrash());
+          wiFiName: secondWifiEntity.name!.getOrCrash(),
+          wiFiPassword: secondWifiEntity.pass!.getOrCrash(),
+        );
       //
       // final SBFirebaseAccountAndSecurityBearSetup
       //     sBFirebaseAccountAndSecurityBearSetup =
@@ -92,9 +96,10 @@ class SecurityBearServerClient {
   }
 
   static ClientChannel createSmartServerClient(String deviceIp) {
-    return ClientChannel(deviceIp,
-        port: 50052,
-        options:
-            const ChannelOptions(credentials: ChannelCredentials.insecure()));
+    return ClientChannel(
+      deviceIp,
+      port: 50052,
+      options: const ChannelOptions(credentials: ChannelCredentials.insecure()),
+    );
   }
 }
