+++ lib/infrastructure/software_info/software_info_repository.dart
@@ -28,6 +28,7 @@ class SoftwareInfoRepository implements ISoftwareInfoRepository {
       final CbjAppIno cbjAppIno = CbjAppIno(
         deviceName: 'cbj App',
         pubspecYamlVersion: packageInfo.version,
+        pubspecYamlBuildNumber: packageInfo.buildNumber,
         protoLastGenDate: appServerProtocGenDate,
       );
